Removes `find.go` file
