+++ internal/service/servicequotas/service_quota.go
@@ -16,6 +16,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/servicequotas/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
@@ -313,4 +314,62 @@ func resourceServiceQuotaParseID(id string) (string, string, error) {
 	}
 
 	return parts[0], parts[1], nil
+}
+
+func findServiceQuotaDefaultByID(ctx context.Context, conn *servicequotas.Client, serviceCode, quotaCode string) (*types.ServiceQuota, error) {
+	input := &servicequotas.GetAWSDefaultServiceQuotaInput{
+		ServiceCode: aws.String(serviceCode),
+		QuotaCode:   aws.String(quotaCode),
+	}
+
+	output, err := conn.GetAWSDefaultServiceQuota(ctx, input)
+
+	if err != nil {
+		return nil, err
+	}
+	if output == nil || output.Quota == nil {
+		return nil, tfresource.NewEmptyResultError(input)
+	}
+
+	return output.Quota, nil
+}
+
+func findServiceQuotaByID(ctx context.Context, conn *servicequotas.Client, serviceCode, quotaCode string) (*types.ServiceQuota, error) {
+	input := &servicequotas.GetServiceQuotaInput{
+		ServiceCode: aws.String(serviceCode),
+		QuotaCode:   aws.String(quotaCode),
+	}
+
+	output, err := conn.GetServiceQuota(ctx, input)
+
+	var nsr *types.NoSuchResourceException
+	if errors.As(err, &nsr) {
+		return nil, &retry.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	if output == nil || output.Quota == nil {
+		return nil, tfresource.NewEmptyResultError(input)
+	}
+
+	if output.Quota.ErrorReason != nil {
+		return nil, &retry.NotFoundError{
+			Message:     fmt.Sprintf("%s: %s", output.Quota.ErrorReason.ErrorCode, aws.ToString(output.Quota.ErrorReason.ErrorMessage)),
+			LastRequest: input,
+		}
+	}
+
+	if output.Quota.Value == nil {
+		return nil, &retry.NotFoundError{
+			Message:     "empty value",
+			LastRequest: input,
+		}
+	}
+
+	return output.Quota, nil
 }
